+++ src/app/src/components/FacilityDetailSidebarItem.jsx
@@ -1,8 +1,10 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { withStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemText from '@material-ui/core/ListItemText';
 
 import FacilityDetailSidebarDetail from './FacilityDetailSidebarDetail';
+import ShowOnly from './ShowOnly';
 
 const detailsSidebarStyles = () =>
     Object.freeze({
@@ -10,10 +12,28 @@ const detailsSidebarStyles = () =>
             paddingTop: '16px',
         },
         label: {
-            padding: '12px 24px 4px 24px',
             fontSize: '0.75rem',
             textTransform: 'uppercase',
             fontWeight: 'bold',
+        },
+        primaryText: {
+            wordWrap: 'break-word',
+        },
+        secondaryText: {
+            color: 'rgba(0, 0, 0, 0.54)',
+            display: 'flex',
+            alignItems: 'center',
+            fontSize: '12px',
+            justify: 'flex-end',
+        },
+        divider: {
+            backgroundColor: 'rgba(0, 0, 0, 0.06)',
+        },
+        icon: {
+            color: 'rgb(106, 106, 106)',
+            fontSize: '24px',
+            fontWeight: 300,
+            textAlign: 'center',
         },
     });
 
@@ -24,23 +44,45 @@ const FacilityDetailSidebarItem = ({
     secondary,
     classes,
     embed,
-    href,
-    link,
+    verified,
 }) => {
-    const hasAdditionalContent = !!additionalContent?.length;
+    const [isOpen, setIsOpen] = useState(false);
+    const hasAdditionalContent = !embed && !!additionalContent?.length;
+    const additionalContentCount = additionalContent?.length;
 
     return (
         <div className={classes.item}>
-            <Typography className={classes.label}>{label}</Typography>
+            <ListItem
+                button={hasAdditionalContent}
+                onClick={() => {
+                    if (!hasAdditionalContent) return;
+                    setIsOpen(!isOpen);
+                }}
+            >
+                <ListItemText
+                    primary={label}
+                    classes={{ primary: classes.label }}
+                />
+                <ShowOnly when={hasAdditionalContent}>
+                    <div className={classes.secondaryText}>
+                        <ListItemText secondary={additionalContentCount + 1} />
+                        <i
+                            className={`${classes.icon} far fa-fw fa-${
+                                isOpen ? 'angle-up' : 'angle-down'
+                            }`}
+                        />
+                    </div>
+                </ShowOnly>
+            </ListItem>
             <FacilityDetailSidebarDetail
-                hasAdditionalContent={!embed && hasAdditionalContent}
-                additionalContentCount={additionalContent?.length}
                 primary={primary}
                 secondary={secondary}
-                onClick={() => {}}
-                link={link}
-                href={href}
+                verified={verified}
             />
+            {isOpen &&
+                additionalContent.map(item => (
+                    <FacilityDetailSidebarDetail {...item} />
+                ))}
         </div>
     );
 };